router: Add initial kernel config
Note that this is just a first-pass smoke test of the system to ensure
that I can boot from my own cat kibble.  A more minimal
router/linux-cfgs/x86_64-targetdisk.cfg may follow in the future, and I
may find I included too little or too much in this kernel config.

The rest of the router image remains to be configured.